Re-organize, and refactor
