+++ src/Emanote/Route/SiteRoute.hs
@@ -16,154 +16,5 @@ module Emanote.Route.SiteRoute
   )
 where
 
-import Control.Lens.Operators ((^.))
-import qualified Data.IxSet.Typed as Ix
-import Data.WorldPeace.Union
-  ( OpenUnion,
-    absurdUnion,
-    openUnionLift,
-  )
-import Ema (Ema (..))
-import qualified Emanote.Model as M
-import qualified Emanote.Model.Note as N
-import qualified Emanote.Model.StaticFile as SF
-import Emanote.Model.Type (Model)
-import Emanote.Prelude (h)
-import Emanote.Route (FileType (Html))
-import qualified Emanote.Route as R
-import Emanote.Route.ModelRoute (LMLRoute, StaticFileRoute)
-
-data IndexR = IndexR
-  deriving (Eq, Show, Ord)
-
-data TagIndexR = TagIndexR
-  deriving (Eq, Show, Ord)
-
--- | A 404 route
-newtype MissingR = MissingR {unMissingR :: FilePath}
-  deriving (Eq, Show, Ord)
-
--- | A route to a virtual resource (not in `Model`)
-type VirtualRoute' =
-  '[ IndexR,
-     TagIndexR
-   ]
-
-type VirtualRoute = OpenUnion VirtualRoute'
-
--- | A route to a resource in `Model`
---
--- This is *mostly isomorphic* to `ModelRoute`, except for containing the
--- absolute path to the static file.
-type ResourceRoute' =
-  '[ (StaticFileRoute, FilePath),
-     LMLRoute
-   ]
-
-type ResourceRoute = OpenUnion ResourceRoute'
-
-type SiteRoute' =
-  '[ VirtualRoute,
-     ResourceRoute,
-     MissingR
-   ]
-
-type SiteRoute = OpenUnion SiteRoute'
-
-instance Ema Model SiteRoute where
-  encodeRoute model =
-    absurdUnion
-      `h` ( \(MissingR _fp) ->
-              error "emanote: attempt to encode a 404 route"
-          )
-      `h` encodeResourceRoute model
-      `h` encodeVirtualRoute
-
-  decodeRoute model fp =
-    fmap openUnionLift (decodeVirtualRoute fp)
-      <|> decodeGeneratedRoute model fp
-      <|> pure (openUnionLift $ MissingR fp)
-
-  allRoutes model =
-    let htmlRoutes =
-          model ^. M.modelNotes
-            & Ix.toList
-            <&> noteFileSiteRoute
-        staticRoutes =
-          model ^. M.modelStaticFiles
-            & Ix.toList
-            <&> staticFileSiteRoute
-        virtualRoutes :: [VirtualRoute] =
-          [openUnionLift IndexR, openUnionLift TagIndexR]
-     in htmlRoutes
-          <> staticRoutes
-          <> fmap openUnionLift virtualRoutes
-
-encodeResourceRoute :: Model -> ResourceRoute -> FilePath
-encodeResourceRoute model =
-  absurdUnion
-    `h` ( \(r :: LMLRoute) ->
-            R.encodeRoute $
-              maybe (coerce . R.lmlRouteCase $ r) N.noteHtmlRoute $
-                M.modelLookupNoteByRoute r model
-        )
-    `h` ( \(r :: StaticFileRoute, _fpAbs :: FilePath) ->
-            R.encodeRoute r
-        )
-
-encodeVirtualRoute :: VirtualRoute -> FilePath
-encodeVirtualRoute =
-  absurdUnion
-    `h` ( \TagIndexR ->
-            R.encodeRoute $ R.mkRouteFromSlug @'Html "@tags"
-        )
-    `h` ( \IndexR ->
-            R.encodeRoute $ R.mkRouteFromSlug @'Html "@index"
-        )
-
--- | Decode a route that is known to refer to a resource in the model
-decodeGeneratedRoute :: Model -> FilePath -> Maybe SiteRoute
-decodeGeneratedRoute model fp =
-  fmap
-    staticFileSiteRoute
-    (flip M.modelLookupStaticFileByRoute model =<< R.decodeAnyRoute fp)
-    <|> fmap
-      noteFileSiteRoute
-      (flip M.modelLookupNoteByHtmlRoute model $ R.decodeHtmlRoute fp)
-
-decodeVirtualRoute :: FilePath -> Maybe VirtualRoute
-decodeVirtualRoute fp =
-  fmap openUnionLift (decodeIndexR fp)
-    <|> fmap openUnionLift (decodeTagIndexR fp)
-
-decodeIndexR :: FilePath -> Maybe IndexR
-decodeIndexR fp = do
-  slug <- R.routeSlug . R.decodeHtmlRoute $ fp
-  guard $ slug == "@index"
-  pure IndexR
-
-decodeTagIndexR :: FilePath -> Maybe TagIndexR
-decodeTagIndexR fp = do
-  slug <- R.routeSlug . R.decodeHtmlRoute $ fp
-  guard $ slug == "@tags"
-  pure TagIndexR
-
-noteFileSiteRoute :: N.Note -> SiteRoute
-noteFileSiteRoute =
-  lmlSiteRoute . N._noteRoute
-
-lmlSiteRoute :: LMLRoute -> SiteRoute
-lmlSiteRoute =
-  openUnionLift . lmlResourceRoute
-
-lmlResourceRoute :: LMLRoute -> ResourceRoute
-lmlResourceRoute =
-  openUnionLift
-
-staticFileSiteRoute :: SF.StaticFile -> SiteRoute
-staticFileSiteRoute =
-  (openUnionLift . staticResourceRoute) . (SF._staticFileRoute &&& SF._staticFilePath)
-  where
-    staticResourceRoute :: (StaticFileRoute, FilePath) -> ResourceRoute
-    staticResourceRoute =
-      openUnionLift
+import Emanote.Route.SiteRoute.Class
+import Emanote.Route.SiteRoute.Type
